requirements.txt
fastapi==0.95.0
librosa==0.9.2
soundfile==0.11.0
matplotlib==3.6.2
numpy==1.23.0
Pillow==9.4.0
pytest==7.2.1
requests==2.25.1
streamlit==1.19.0
uvicorn==0.15.0
